requirements.txt
google-generativeai
pandas
tabula-py
pytest
python-dotenv